+++ server/libs/platform/platform-configuration/platform-configuration-rest/platform-configuration-rest-api/src/main/java/com/bytechef/platform/configuration/web/rest/mapper/ScriptExecutionMapper.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2023-present ByteChef Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.platform.configuration.web.rest.mapper;
+
+import com.bytechef.platform.configuration.dto.ScriptTestExecutionDTO;
+import com.bytechef.platform.configuration.web.rest.mapper.config.PlatformConfigurationMapperSpringConfig;
+import com.bytechef.platform.configuration.web.rest.model.ScriptTestExecutionModel;
+import org.mapstruct.Mapper;
+import org.springframework.core.convert.converter.Converter;
+
+/**
+ * @author Ivica Cardic
+ */
+@Mapper(config = PlatformConfigurationMapperSpringConfig.class)
+public interface ScriptExecutionMapper extends Converter<ScriptTestExecutionDTO, ScriptTestExecutionModel> {
+
+    @Override
+    ScriptTestExecutionModel convert(ScriptTestExecutionDTO scriptTestExecutionDTO);
+}
